backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9
python-decouple==3.8
Pillow>=10.0.0
openpyxl>=3.1.0

